rename base script to base.py, change associated scripts accordingly
